split filesystementry up
